pyproject.toml
[tool.poetry]
name = "hw7"
version = "0.1.0"
description = ""
authors = ["Oleg Smirnov <oleg.smirnoff505@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.1"
pillow = "^10.2.0"
django-filter = "^23.5"
ipython = "^8.21.0"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
stripe = "^8.4.0"
celery = "^5.3.6"
redis = "^5.0.2"
django-celery-beat = "^2.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
